Update import paths for hooks in various components
