+++ app/hooks/admin/useCategoriesWithPagination.ts
@@ -1,15 +1,25 @@
-import { getCategories } from '../lib/api/shop/category'
-import { useToast } from '../components/ui/use-toast'
+import { getCategories } from '../../lib/api/shop/category'
+import { useToast } from '../../components/ui/use-toast'
 import { useEffect, useState } from 'react'
 
 import type { GetAllOfModel } from '@/types/shared/getAllOfModel'
 import type { Category } from '@/types/shop/category.types'
 
-const useCategories = ({
-  isAvailable, limit, page
-}: GetAllOfModel): { categories: Category[], isLoading: boolean } => {
+interface UseCategoriesWithPaginationResponse {
+  page: number
+  total: number
+  setPage: (page: number) => void
+  isLoading: boolean
+  categories: Category[]
+}
+
+const useCategoriesWithPagination = ({
+  isAvailable, limit
+}: GetAllOfModel): UseCategoriesWithPaginationResponse => {
   const [categories, setCategories] = useState<Category[]>([])
   const [isLoading, setIsLoading] = useState<boolean>(true)
+  const [total, setTotal] = useState<number>(0)
+  const [page, setPage] = useState<number>(1)
   const { toast } = useToast()
 
   useEffect(() => {
@@ -17,6 +27,7 @@ const useCategories = ({
       try {
         const categories = await getCategories({ isAvailable, limit, page })
         setCategories(categories.categories)
+        setTotal(categories.total)
         setIsLoading(false)
       } catch (error) {
         toast({
@@ -27,12 +38,15 @@ const useCategories = ({
     }
 
     void fetchCategories()
-  }, [])
+  }, [limit, page])
 
   return {
-    categories,
-    isLoading
+    page,
+    total,
+    setPage,
+    isLoading,
+    categories
   }
 }
 
-export default useCategories
+export default useCategoriesWithPagination
