+++ app/hooks/shop/useCartButtonValidation.ts
