+++ app/hooks/shop/useCategories.ts
@@ -1,25 +1,15 @@
-import { getCategories } from '../lib/api/shop/category'
-import { useToast } from '../components/ui/use-toast'
+import { getCategories } from '../../lib/api/shop/category'
+import { useToast } from '../../components/ui/use-toast'
 import { useEffect, useState } from 'react'
 
 import type { GetAllOfModel } from '@/types/shared/getAllOfModel'
 import type { Category } from '@/types/shop/category.types'
 
-interface UseCategoriesWithPaginationResponse {
-  page: number
-  total: number
-  setPage: (page: number) => void
-  isLoading: boolean
-  categories: Category[]
-}
-
-const useCategoriesWithPagination = ({
-  isAvailable, limit
-}: GetAllOfModel): UseCategoriesWithPaginationResponse => {
+const useCategories = ({
+  isAvailable, limit, page
+}: GetAllOfModel): { categories: Category[], isLoading: boolean } => {
   const [categories, setCategories] = useState<Category[]>([])
   const [isLoading, setIsLoading] = useState<boolean>(true)
-  const [total, setTotal] = useState<number>(0)
-  const [page, setPage] = useState<number>(1)
   const { toast } = useToast()
 
   useEffect(() => {
@@ -27,7 +17,6 @@ const useCategoriesWithPagination = ({
       try {
         const categories = await getCategories({ isAvailable, limit, page })
         setCategories(categories.categories)
-        setTotal(categories.total)
         setIsLoading(false)
       } catch (error) {
         toast({
@@ -38,15 +27,12 @@ const useCategoriesWithPagination = ({
     }
 
     void fetchCategories()
-  }, [limit, page])
+  }, [])
 
   return {
-    page,
-    total,
-    setPage,
-    isLoading,
-    categories
+    categories,
+    isLoading
   }
 }
 
-export default useCategoriesWithPagination
+export default useCategories
